feature:最终bug调整
