+++ web/package.json
@@ -20,7 +20,7 @@
         "axios": "^1.4.0",
         "core-js": "^3.31.1",
         "echarts": "5.4.3",
-        "element-plus": "^2.3.8",
+        "element-plus": "^2.7.0",
         "highlight.js": "^11.8.0",
         "js-cookie": "^3.0.5",
         "marked": "4.3.0",
