+++ web/src/components/commandMenu/index.vue
@@ -41,17 +41,17 @@
 
 <script setup>
 import { reactive, ref, watch } from 'vue'
-import { useRouter, useRoute } from 'vue-router'
+import { useRouter } from 'vue-router'
 import { useRouterStore } from '@/pinia/modules/router'
-import { useUserStore } from '@/pinia/modules/user'
-
+import { useAppStore,useUserStore } from '@/pinia'
 defineOptions({
   name: 'CommandMenu',
 })
+const appStore = useAppStore()
+const userStore = useUserStore()
 
 const router = useRouter()
 const route = useRouter()
-const userStore = useUserStore()
 const routerStore = useRouterStore()
 const dialogVisible = ref(false)
 const searchInput = ref('')
@@ -134,10 +134,10 @@ const changeRouter = (e) => {
 
 const changeMode = (e) => {
   if (e === null) {
-    userStore.changeSideMode('dark')
+    appStore.toggleTheme(false )
     return
   }
-  userStore.changeSideMode(e)
+  appStore.toggleTheme(true )
 }
 
 const close = () => {
@@ -172,6 +172,7 @@ watch(searchInput, () => {
     color: #666;
   }
   .quick-input{
+    @apply bg-gray-50 dark:bg-gray-800;
     color: #666;
     border-radius: 4px 4px 0 0;
     border:none;
@@ -186,8 +187,8 @@ watch(searchInput, () => {
     padding: 8px;
     margin: 4px 0;
     &:hover{
+      @apply bg-gray-200 dark:bg-slate-500;
       cursor: pointer;
-      background: #eee;
       border-radius: 4px;
     }
   }
